feat: Upgrade parsers and tools
Use the latest versions of everything. Replace package-size
with package-cost. The former is not maintained any more.

BREAKING CHANGE: Sizes are computed differently by package-cost,
which tries to build an application with the dependency to be able
to measure all bundles files, sucluding sub-dependencies. The new
computed sizes may differ, but generally they should correlate
with the earlier results.